Revert "Fixed bug #75961 (Strange references behavior)"
This reverts commit f12662ed6854d62b701bf75a3f4a30ea97a1ba92.

This code needs to be mindful about modifications to the array
happening during callback execution. It was written in a way that
only accessed the reference, which is guaranteed not to move. The
changed implementation instead accesses the array slot, leading to
use-after-free.

Run ext/standard/tests/array/bug61967.phpt under valgrind to see
the issue.